+++ src/lib/UserTasksPopover.svelte
@@ -8,18 +8,21 @@
   import PinAngleIcon from 'svelte-bootstrap-icons/lib/PinAngleFill.svelte';
   import { addTask } from "@stores/todoListStore.js";
   import { userStore } from "@stores/userStore.js";
+  import { isMobileWidth } from "@utils/helper.js";
 
   export let trigger;
   export let user;
+  export let quest;
   export let visible = false;
   
   const profileService = new ProfilesService();
+  const isMobile = isMobileWidth();
 
   let element;
   let parent;
   let arrowElement;
   let tasks = [];
-  let loading = true;
+  let loading = false;
   let canAddTasks = false;
 
   const dispatchEvent = createEventDispatcher();
@@ -29,28 +32,36 @@
       user.sessionId,
       user.questId
     );
-    tasks = response.tasks;
+    tasks = (response.tasks || []).map(task => {
+      task.quest = quest;
+      return task;
+    })
   }
 
-  function showPopover() {
+  function showPopover(evt) {
+    fetchContent();
     visible = true;
     document.addEventListener("click", handleClickOutside);
     window.addEventListener("resize", positionPopover);
+    if (isMobile) {
+      const bdrop = document.createElement("div");
+      bdrop.classList.add("custom-popover-backdrop");
+      document.body.appendChild(bdrop);
+    }
   }
 
-  function hidePopover() {
+  function hidePopover(evt) {
     visible = false;
     document.removeEventListener("click", handleClickOutside);
     window.removeEventListener("resize", positionPopover);
+    if (isMobile) {
+      document.querySelector(".custom-popover-backdrop")?.remove();
+    }
   }
 
   $: canAddTasks = $userStore?.profileId === user.user;
   $: if (visible && trigger && element) {
-    if (loading) {
-      fetchContent();
-      loading = false;
-    }
-    positionPopover();
+      positionPopover();
   }
 
   function handleClickOutside(event) {
@@ -74,18 +85,18 @@
     const viewportHeight = window.innerHeight;
     const popoverWidth = element.offsetWidth;
     const popoverHeight = element.offsetHeight;
-
+    
     let top = triggerRect.top - parentRect.top;
     let left = triggerRect.left - parentRect.left;
     let arrowVerticalSide = "bottom";
     let arrowHorizontalSide = "right";
     let arrowLeft = 20;
-
+    
     const hasRoomBelow = triggerRect.bottom + popoverHeight <= viewportHeight;
     const hasRoomAbove = triggerRect.top - popoverHeight >= 0;
     const hasRoomRight = triggerRect.right + popoverWidth <= viewportWidth;
     const hasRoomLeft = triggerRect.left - popoverWidth >= 0;
-
+    ;
     if (hasRoomBelow) {
       top += triggerRect.height + 5;
       arrowVerticalSide = "bottom";
@@ -93,11 +104,12 @@
       top -= popoverHeight + 5;
       arrowVerticalSide = "top";
     }
-
-    if (!hasRoomRight) {
-      left -= popoverWidth - triggerRect.width;
+    if (!hasRoomRight && hasRoomLeft) {
+      left -= popoverWidth - triggerRect.width - 10;
       arrowHorizontalSide = "left";
       arrowLeft = popoverWidth - 40;
+    } else if (!hasRoomLeft) {
+      left = 0;
     }
 
     element.style.top = `${top}px`;
@@ -118,12 +130,13 @@
 
   onMount(() => {
     trigger.addEventListener("click", showPopover);
-    parent = trigger.closest(".card");
+    parent = trigger.closest(".card-quest");
 
     return () => {
       document.removeEventListener("click", handleClickOutside);
       trigger.removeEventListener("click", showPopover);
       window.removeEventListener("resize", positionPopover);
+      document.querySelector(".custom-popover-backdrop")?.remove();
     };
   });
 
@@ -135,13 +148,6 @@
     });
     hidePopover();
   }
-
-  function onPinTask(event) {
-    const task = event.detail;
-    if (!task.accomplished) {
-      addTask(task);
-    }
-  }
 </script>
 
 {#if visible}
@@ -149,7 +155,7 @@
     class="custom-popover"
     transition:fade={{ duration: 200 }}
     bind:this={element}
-  >
+  >   
     <div class="custom-popover-arrow" bind:this={arrowElement}></div>
     <h3 class="custom-popover-header">
       <span>{user.user}'s Tasks</span>
@@ -159,7 +165,7 @@
         {/if}
     </h3>
     <div class="custom-popover-body">
-      <UserTasks {tasks} on:change={onChangeUserTasks} on:pin={onPinTask} {canAddTasks}/>      
+      <UserTasks {tasks} on:change={onChangeUserTasks} {canAddTasks}/>      
     </div>
   </div>
 {/if}