+++ src/services/AuthService.js
@@ -3,6 +3,8 @@ import { get } from 'svelte/store';
 import { jwtStore, refreshTokenStore } from '../stores/jwtStore';
 import { userStore } from '../stores/userStore';
 import { NetworkService } from './NetworkService';
+import Bowser from "bowser";
+import { version } from 'pino';
 
 const apiUrl = import.meta.env.VITE_API_URL;
 
@@ -23,18 +25,46 @@ export class AuthService {
     return roles.some((role) => u.roles.includes(role));
   }
 
+  _getDeviceInfo() {
+    const browser = Bowser.getParser(window.navigator.userAgent);
+    return {
+      browser: browser.getBrowserName(),
+      os: browser.getOSName(),
+      device: browser.getPlatformType(),
+      version: browser.getBrowserVersion(),
+      language: navigator.language,
+      timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
+    };
+  }
+
+  _getDeviceId() {
+    let deviceId =  localStorage.getItem('uuid');
+    if (!deviceId) {
+      deviceId = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+        var r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8);
+        return v.toString(16);
+      });
+      localStorage.setItem('uuid', deviceId);
+    }
+    return deviceId;
+  }
+
   async logout() {
     await this.networkService.get({
       uri: '/api/logout',
       auth: NetworkService.AUTH_JWT
     });
+    localStorage.removeItem('uuid');
     this.removePersistence();
   }
 
   async login(username, password) {
+    const deviceInfo = this._getDeviceInfo();
+    const deviceId = this._getDeviceId();
+    
     const res = await this.networkService.post({
       uri: '/api/login',
-      body: JSON.stringify({ username, password }),
+      body: JSON.stringify({ username, password, deviceInfo, deviceId }),
       auth: NetworkService.AUTH_NONE,
     });