+++ src/utilities/CustomPopover.js
@@ -1,0 +1,25 @@
+export class CustomPopover {
+  constructor({
+    trigger,
+    element
+  }) {
+    this.trigger = trigger;
+    this.element = element;
+    this.isVisible = false;
+  }
+
+  show() {
+
+  }
+
+  hide() {
+
+  }
+
+  handleClickOutside() {
+
+  }
+
+  getPosition() {
+  }
+}