+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [2.6.3] - 2025-01-24
+### Changed
+- Todo
+
+## [2.6.2] - 2025-01-23
+### Changed
+- Mobile task list revision 1
+- Popover support for mobile
+
 ## [2.6.1] - 2025-01-22
 ### Changed
 - Added light mode for new popovers
@@ -58,7 +67,9 @@ All notable changes to this project will be documented in this file.
 ## [2.0.2] - 2024-06-15
 Initial commit
 
-[Unreleased]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.6.1...HEAD
+[Unreleased]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.6.3...HEAD
+[2.6.3]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.6.2...2.6.3
+[2.6.2]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.6.1...2.6.2
 [2.6.1]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.6.0...2.6.1
 [2.6.0]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.5.1...2.6.0
 [2.5.1]: https://github.com/jmconde/SPTProfileConnectWeb/compare/2.5.0...2.5.1